Pull out adding repo code
